Move HMH to wip/hmh.h, as I don't currently trust it.
