+++ testsrc/hmhtest.cpp
@@ -1,4 +1,5 @@
 #include "mh.h"
+#include "wip/hmh.h"
 #include <unordered_set>
 #include <random>
